requirements.txt
# Python dependencies
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-dotenv==1.0.0
pydantic==2.5.3
youtube-transcript-api==0.6.2
langchain==0.1.5
langchain-community==0.0.16
langchain-cohere==0.0.4
langchain-groq==0.0.1
faiss-cpu==1.7.4
jinja2==3.1.3
aiofiles==23.2.1
python-multipart
